CRA vaihto Viteen, koska CRA ei päivity ja valittaa yhteensopivuuksista typescriptissä
