+++ elokuvaprojekti/server/models/Favourite.js
@@ -1,10 +1,12 @@
-const { pool } = require('../helper/db.js');
+import { pool } from '../helper/db.js';
 
 // Käyttäjän suosikit
 class Favourite {
     static async add(userId, tmdbId) {
         await pool.query(
-            'INSERT INTO favourites (userid, tmdbid) VALUES ($1, $2) ON CONFLICT DO NOTHING',
+            `INSERT INTO favourites (userid, tmdbid) 
+            VALUES ($1, $2) 
+            ON CONFLICT DO NOTHING`,
             [userId, tmdbId]
         );
     }
@@ -12,7 +14,9 @@ class Favourite {
     // Poistaa elokuvan käyttäjän suosikeista
     static async remove(userId, tmdbId) {
         await pool.query(
-            'DELETE FROM favourites WHERE userid = $1 AND tmdbid = $2',
+            `DELETE FROM favourites 
+            WHERE userid = $1 
+            AND tmdbid = $2`,
             [userId, tmdbId]
         );
     }
@@ -20,11 +24,13 @@ class Favourite {
     // Hakee käyttäjän suosikit
     static async getByUser(userId) {
         const result = await pool.query(
-            'SELECT tmdbid FROM favourites WHERE userid = $1',
+            `SELECT tmdbid 
+            FROM favourites 
+            WHERE userid = $1`,
             [userId]
         );
         return result.rows;
     }
 }
 
-module.exports = Favourite;
+export default Favourite;
