+++ elokuvaprojekti/src/api/groupmovies.jsx
@@ -1,6 +1,6 @@
 import axios from "axios";
 
-const API_URL = process.env.REACT_APP_API_URL;
+const API_URL = import.meta.env.VITE_API_URL || "http://localhost:3001";
 
 // Lisää uusi elokuvajako
 export const addGroupMovie = async (payload, token) => {
