+++ elokuvaprojekti/src/api/moviedb.jsx
@@ -1,4 +1,6 @@
-const BASE_URL = process.env.REACT_APP_API_URL || "http://localhost:3001"
+const BASE_URL = import.meta.env.VITE_API_URL || "http://localhost:3001";
+
+export default BASE_URL;
 
 // --- TMDB API-rajapintafunktiot (Backendin kautta) --- //
