+++ elokuvaprojekti/src/components/FavouriteList.jsx
@@ -1,6 +1,6 @@
 import { useParams, Link } from 'react-router-dom';
 import { useEffect, useState } from 'react';
-import { getMovieDetails } from '../api/moviedb';
+import { getMovieDetails } from '../api/moviedb.jsx';
 
 function FavouriteList() {
     const { userId } = useParams();
