+++ elokuvaprojekti/src/components/GroupMovieCard.jsx
