+++ elokuvaprojekti/src/components/GroupShowCard.jsx
