+++ elokuvaprojekti/src/components/MovieCard.jsx
