+++ elokuvaprojekti/src/components/ProtectedRoute.jsx
@@ -1,5 +1,5 @@
 import  { useContext } from 'react'
-import { AuthContext } from '../context/authContext.js'
+import { AuthContext } from '../context/authContext.jsx'
 import { Navigate } from 'react-router-dom'
 
 export default function ProtectedRoute({ children }) {
