+++ elokuvaprojekti/src/components/ReviewCard.jsx
