+++ elokuvaprojekti/src/components/ShareShowModal.jsx
@@ -1,18 +1,21 @@
-import React, { useState, useEffect } from "react";
-import { addGroupMovie } from "../api/groupmovies";
-import axios from "axios";
+import React, { useState, useEffect } from 'react';
+import axios from 'axios';
+import { useNavigate } from 'react-router-dom';
 
-const ShareMovieModal = ({ onClose, movieData, onShared }) => {
+const ShareShowModal = ({ onClose, showData, onShared }) => {
   const [groups, setGroups] = useState([]);
   const [formData, setFormData] = useState({
-    groupID: "",
-    reason: ""
+    groupID: '',
+    reason: ''
   });
-  const [error, setError] = useState("");
+  const [error, setError] = useState('');
   const [hasToken, setHasToken] = useState(false);
 
+  const navigate = useNavigate();
+
+  // Haetaan kirjautuneen käyttäjän ryhmät
   useEffect(() => {
-    const token = localStorage.getItem("token");
+    const token = localStorage.getItem('token');
     if (!token) {
       setHasToken(false);
       setGroups([]);
@@ -23,13 +26,13 @@ const ShareMovieModal = ({ onClose, movieData, onShared }) => {
     const fetchGroups = async () => {
       try {
         const res = await axios.get(
-          `${process.env.REACT_APP_API_URL}/groups/mine`,
+          `${import.meta.env.VITE_API_URL}/groups/mine`,
           { headers: { Authorization: `Bearer ${token}` } }
         );
         setGroups(res.data);
       } catch (err) {
-        console.error("Virhe ryhmien haussa:", err);
-        setError("Ryhmiä ei voitu hakea");
+        console.error('Virhe ryhmien haussa:', err);
+        setError('Ryhmiä ei voitu hakea');
       }
     };
 
@@ -42,31 +45,40 @@ const ShareMovieModal = ({ onClose, movieData, onShared }) => {
 
   const handleSubmit = async (e) => {
     e.preventDefault();
-    setError("");
+    setError('');
 
-    const token = localStorage.getItem("token");
+    const token = localStorage.getItem('token');
     if (!token) {
-      setError("Sinun täytyy kirjautua sisään jakaaksesi elokuvan.");
+      setError("Sinun täytyy kirjautua sisään jakaaksesi näytöksen.");
       return;
     }
 
     try {
       const payload = {
         groupID: formData.groupID,
-        tmdbID: movieData.tmdbId,
-        movieName: movieData.name,
-        image: movieData.image || null,
-        url: movieData.url || null,
-        reason: formData.reason
+        showID: showData.id,
+        eventID: showData.eventId,
+        tmdbID: showData.tmdbId || null,
+        theatre: showData.theatre,
+        auditorium: showData.auditorium,
+        showTime: showData.time,
+        reason: formData.reason,
+        image: showData.image || null,
+        url: showData.url || null,
+        movieName: showData.name || null
       };
 
-      const newShare = await addGroupMovie(payload, token);  // 👈 API-kutsu
+      const res = await axios.post(
+        `${import.meta.env.VITE_API_URL}/groupshows`,
+        payload,
+        { headers: { Authorization: `Bearer ${token}` } }
+      );
 
-      if (onShared) onShared(newShare);
+      if (onShared) onShared(res.data);
       onClose();
     } catch (err) {
-      console.error("Virhe jaettaessa elokuvaa:", err.response?.data || err.message);
-      setError(err.response?.data?.message || "Elokuvan jako epäonnistui");
+      console.error('Virhe jaettaessa näytöstä:', err.response?.data || err.message);
+      setError(err.response?.data?.message || 'Näytöksen jako epäonnistui');
     }
   };
 
@@ -75,7 +87,7 @@ const ShareMovieModal = ({ onClose, movieData, onShared }) => {
       <div className="modal-dialog modal-dialog-centered" role="document">
         <div className="modal-content rounded-xl">
           <div className="modal-header">
-            <h5 className="modal-title">Jaa elokuva ryhmälle</h5>
+            <h5 className="modal-title">Jaa näytös ryhmälle</h5>
             <button
               type="button"
               className="btn-close"
@@ -84,24 +96,24 @@ const ShareMovieModal = ({ onClose, movieData, onShared }) => {
             ></button>
           </div>
           <div className="modal-body">
-            <div className="mb-3 text-center">
-              {movieData.image && (
-                <img
-                  src={movieData.image}
-                  alt={movieData.name}
-                  style={{ maxWidth: "150px", borderRadius: "10px" }}
-                  className="mb-2"
-                />
-              )}
-              <div><strong>{movieData.name}</strong></div>
+            {/* Näytöksen tiedot */}
+            <div className="mb-3">
+              <strong>{showData.name}</strong><br />
+              {showData.theatre}{showData.auditorium ? `, ${showData.auditorium}` : ''}<br />
+              {new Date(showData.time).toLocaleString('fi-FI', {
+                dateStyle: 'short',
+                timeStyle: 'short'
+              })}
             </div>
 
+            {/* Jos ei ole kirjautunut */}
             {!hasToken ? (
               <div className="alert alert-warning text-center">
-                <p>Kirjaudu sisään jakaaksesi elokuvan.</p>
+                <p>Kirjaudu sisään jakaaksesi näytöksen.</p>
               </div>
             ) : (
               <form onSubmit={handleSubmit}>
+                {/* Dropdown ryhmille */}
                 <div className="mb-3">
                   <label htmlFor="groupID" className="form-label">Valitse ryhmä</label>
                   <select
@@ -121,8 +133,9 @@ const ShareMovieModal = ({ onClose, movieData, onShared }) => {
                   </select>
                 </div>
 
+                {/* Perusteluteksti */}
                 <div className="mb-3">
-                  <label htmlFor="reason" className="form-label">Miksi jaat tämän elokuvan?</label>
+                  <label htmlFor="reason" className="form-label">Miksi jaat tämän näytöksen?</label>
                   <textarea
                     name="reason"
                     id="reason"
@@ -135,7 +148,7 @@ const ShareMovieModal = ({ onClose, movieData, onShared }) => {
                 </div>
 
                 <button type="submit" className="btn btn-primary w-100 mt-2">
-                  Jaa elokuva
+                  Jaa näytös
                 </button>
                 {error && <p className="text-danger mt-3">{error}</p>}
               </form>
@@ -147,4 +160,4 @@ const ShareMovieModal = ({ onClose, movieData, onShared }) => {
   );
 };
 
-export default ShareMovieModal;
+export default ShareShowModal;
